[timeseries] reorganize gluonts namespace, fix minor typing issue in metrics
