+++ timeseries/src/autogluon/timeseries/models/chronos/pipeline/utils.py
@@ -16,7 +16,7 @@ from transformers import TrainerCallback
 from autogluon.common.loaders.load_s3 import download, list_bucket_prefix_suffix_contains_s3
 from autogluon.core.utils.exceptions import TimeLimitExceeded
 from autogluon.timeseries.dataset.ts_dataframe import TimeSeriesDataFrame
-from autogluon.timeseries.models.gluonts.abstract_gluonts import SimpleGluonTSDataset
+from autogluon.timeseries.models.gluonts.abstract import SimpleGluonTSDataset
 
 if TYPE_CHECKING:
     # TODO: fix the underlying reason for this circular import, the pipeline should handle tokenization
