+++ timeseries/tests/unittests/test_metrics.py
@@ -27,7 +27,7 @@ from autogluon.timeseries.metrics import (
     check_get_evaluation_metric,
 )
 from autogluon.timeseries.metrics.utils import in_sample_abs_seasonal_error, in_sample_squared_seasonal_error
-from autogluon.timeseries.models.gluonts.abstract_gluonts import AbstractGluonTSModel
+from autogluon.timeseries.models.gluonts.abstract import AbstractGluonTSModel
 
 from .common import DUMMY_TS_DATAFRAME, get_data_frame_with_item_index, get_prediction_for_df
